+++ src/components/v2/MultiRangeSlider/RangeSlider.jsx
