+++ src/components/v2/games/carousel.jsx
