+++ src/components/v2/games/details.jsx
