+++ src/components/v2/games/difficulty.jsx
