+++ src/components/v2/games/rating.jsx
