+++ src/components/v2/home/hotpicks.jsx
