+++ src/pages/v2/news/[id]/index.tsx
